+++ src/version/index.ts
@@ -1,4 +1,4 @@
-export const VERSION = '5.0.23';
+export const VERSION = '5.0.24';
 
 export function getVersion(): string {
   return VERSION;